gps: Move IgnoredRuleset into its own file
